Phase 0 submission
The app works.  The window color is set according to the coordinates at
which MouseEvents occur (specifically mouse down events).  The blue
component is determined by a random number generator.

The previous commit errors seem to be caused by a misnamed repository
inside this one.  That has been addressed.

Further comments on how this app works can be found in the HelloWorldApp
repository from which this was taken.